+++ config/importmap.rb
@@ -2,3 +2,4 @@ pin_all_from File.expand_path("../app/assets/javascripts", __dir__)
 pin "@hotwired/turbo-rails", to: "turbo.min.js", preload: true
 pin "@hotwired/stimulus", to: "https://ga.jspm.io/npm:@hotwired/stimulus@3.0.1/dist/stimulus.js"
 pin "@hotwired/stimulus-loading", to: "stimulus-loading.js", preload: true
+pin "showdown", to: "https://ga.jspm.io/npm:showdown@2.1.0/dist/showdown.js"