+++ lib/eyeloupe.rb
@@ -3,11 +3,13 @@ require "eyeloupe/engine"
 
 require 'eyeloupe/request_middleware'
 require 'eyeloupe/configuration'
-require 'eyeloupe/http'
 require 'eyeloupe/processors/in_request'
 require 'eyeloupe/processors/out_request'
+require 'eyeloupe/processors/exception'
+require 'eyeloupe/concerns/rescuable'
 
 require 'pagy'
+require "openai"
 module Eyeloupe
 
   # @return [Eyeloupe::Configuration]