+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-## 0.1.0
+## 0.3.0
 
-- Initial release including incoming and outgoing requests.
+- Add exceptions: Framework + ActiveJob + Sidekiq Worker exceptions.
+- Add OpenAI support for AI assistant in exceptions.
+- Fix exceptions when using importmap binary by adding net/http override in railties initializer.
 
 ## 0.2.0
 
 - Fix missing require for `pagy` gem.
+
+## 0.1.0
+
+- Initial release including incoming and outgoing requests.